Chore: Update pre-commit dependencies to latest
ERROR! the role 'mono-install' was not found in <>
The error appears to be in '': line 5, column 7, but may
be elsewhere in the file depending on the exact syntax
problem.

Update role name to match the role path on GHA and
Jenkins jobs.

Add GHA verify workflow for verify jobs.
Update tox and pre-commit dependencies.

Ref: https://github.com/ansible/molecule/issues/1042
Change-Id: Ifc40c99516a92a1525d2b088a30d7efe34b3b17d
Signed-off-by: Anil Belur <abelur@linuxfoundation.org>